Typ lyckats merga. Allt utom leftcam blinkar, vet inte varför. Bytte temporärt ut skybox texturerna mot grass.png eftersom up/down/etc. inte fanns i git. Nästa steg är att få bort blinkandet. Övriga TODOs: Clear + flatten knapparna reflekteras inte genom parameterändringar
Anttweakbar vägrar commitas, langar en kopia och en readme i Drive

+++ source/text_utils/text2D.cpp
@@ -43,7 +43,7 @@ void initText2D(const char * texturePath){
 
 void printText2D(const char * text, int x, int y, int size){
 
-	unsigned int length = strlen(text);
+	unsigned int length = (int)strlen(text);
 
 	// Fill buffers
 	std::vector<glm::vec2> vertices;
@@ -107,7 +107,7 @@ void printText2D(const char * text, int x, int y, int size){
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
 	// Draw call
-	glDrawArrays(GL_TRIANGLES, 0, vertices.size() );
+	glDrawArrays(GL_TRIANGLES, 0, (int)vertices.size() );
 
 	glDisable(GL_BLEND);
